Use namespaced FileUtils module instead of patching File
