+++ spec/integration/template_dir_spec.rb
@@ -13,7 +13,7 @@ describe 'template directory' do
       master_hook = File.join(hooks_dir, 'overcommit-hook')
       File.exist?(master_hook).should == true
       File.size?(master_hook).should > 0
-      File.symlink?(master_hook).should == false
+      Overcommit::Utils::FileUtils.symlink?(master_hook).should == false
     end
 
     it 'contains all other hooks as symlinks to the master hook' do
@@ -24,7 +24,7 @@ describe 'template directory' do
       end
 
       Overcommit::Utils.supported_hook_types.each do |hook_type|
-        File.symlink?(File.join(hooks_dir, hook_type)).should == true
+        Overcommit::Utils::FileUtils.symlink?(File.join(hooks_dir, hook_type)).should == true
       end
     end
   end
