+++ spec/overcommit/hook_context/pre_commit_spec.rb
@@ -201,7 +201,7 @@ describe Overcommit::HookContext::PreCommit do
     context 'when a broken symlink is staged' do
       around do |example|
         repo do
-          File.symlink('non-existent-file', 'symlink')
+          Overcommit::Utils::FileUtils.symlink('non-existent-file', 'symlink')
           `git add symlink`
           example.run
         end
